패키지 구조 변경, JWT 적용, Exception 세분화

+++ src/main/java/com/youngman/mop/domain/member/service/MemberDeleteService.java
@@ -1,5 +1,6 @@
 package com.youngman.mop.domain.member.service;
 
+import com.youngman.mop.domain.member.exception.MemberNotFoundException;
 import com.youngman.mop.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,11 +15,16 @@ public class MemberDeleteService {
 
 	private final MemberRepository memberRepository;
 
+
 	public void deleteById(Long id) {
 		memberRepository.deleteById(id);
 	}
 
 	public void deleteByEmail(String email) {
-		memberRepository.deleteByEmail(email);
+		if (memberRepository.existsByEmail(email)) {
+			memberRepository.deleteByEmail(email);
+			return;
+		}
+		throw new MemberNotFoundException();
 	}
 }
